@@ -9,37 +9,37 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "albertcht/invisible-recaptcha": "^1.9",
-        "arcanedev/log-viewer": "7.x",
-        "darkghosthunter/laraguard": "^2.0",
+        "php": "^7.3",
+        "arcanedev/log-viewer": "8.x",
+        "darkghosthunter/laraguard": "dev-master",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^2.0",
-        "guzzlehttp/guzzle": "^6.3",
+        "guzzlehttp/guzzle": "^7.0.1",
         "jamesmills/laravel-timezone": "^1.9",
         "lab404/laravel-impersonate": "^1.6",
         "langleyfoxall/laravel-nist-password-rules": "^4.1",
-        "laravel/framework": "^7.24",
-        "laravel/socialite": "^4.3",
+        "laravel/framework": "^8.0",
+        "laravel/socialite": "^5.0",
         "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "livewire/livewire": "^1.3",
-        "rappasoft/laravel-livewire-tables": "^0.1",
-        "rappasoft/lockout": "^2.1",
+        "laravel/ui": "^3.0",
+        "livewire/livewire": "^2.0",
+        "rappasoft/laravel-livewire-tables": "^0.3",
+        "rappasoft/lockout": "^3.0",
         "spatie/laravel-activitylog": "^3.14",
         "spatie/laravel-permission": "^3.11",
-        "tabuna/breadcrumbs": "^1.0"
+        "tabuna/breadcrumbs": "^2.2"
     },
     "require-dev": {
+        "roave/security-advisories": "dev-master",
         "barryvdh/laravel-debugbar": "^3.2",
         "barryvdh/laravel-ide-helper": "^2.6",
-        "codedungeon/phpunit-result-printer": "^0.26.2",
-        "facade/ignition": "^2.0",
+        "codedungeon/phpunit-result-printer": "^0.29",
+        "facade/ignition": "^2.3.6",
         "friendsofphp/php-cs-fixer": "^2.16",
         "fzaninotto/faker": "^1.9.1",
         "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -53,12 +53,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
