@@ -76,7 +76,7 @@ class CreateRoleTest extends TestCase
     /** @test */
     public function only_admin_can_create_roles()
     {
-        $this->actingAs(factory(User::class)->state('admin')->create());
+        $this->actingAs(User::factory()->admin()->create());
 
         $response = $this->get('/admin/auth/role/create');
 
