@@ -24,7 +24,7 @@ class ListRoleTest extends TestCase
     /** @test */
     public function only_admin_can_view_roles()
     {
-        $this->actingAs(factory(User::class)->state('admin')->create());
+        $this->actingAs(User::factory()->admin()->create());
 
         $response = $this->get('/admin/auth/role');
 
