@@ -13,7 +13,7 @@ class ConfirmationTest extends TestCase
     /** @test */
     public function a_user_can_access_the_confirm_password_page()
     {
-        $user = factory(User::class)->create();
+        $user = User::factory()->create();
 
         $this->actingAs($user);
 
