@@ -40,7 +40,7 @@ class RoleEventListener
                     'type' => $event->role->type,
                     'name' => $event->role->name,
                 ],
-                'permissions' =>$event->role->permissions->count() ? $event->role->permissions->pluck('description')->implode(', ') : 'None',
+                'permissions' => $event->role->permissions->count() ? $event->role->permissions->pluck('description')->implode(', ') : 'None',
             ])
             ->log(':causer.name updated role :subject.name with permissions: :properties.permissions');
     }
