@@ -4,6 +4,8 @@ namespace App\Domains\Auth\Models;
 
 use App\Domains\Auth\Models\Traits\Attribute\RoleAttribute;
 use App\Domains\Auth\Models\Traits\Method\RoleMethod;
+use Database\Factories\RoleFactory;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Spatie\Permission\Models\Role as SpatieRole;
 
 /**
@@ -11,7 +13,8 @@ use Spatie\Permission\Models\Role as SpatieRole;
  */
 class Role extends SpatieRole
 {
-    use RoleAttribute,
+    use HasFactory,
+        RoleAttribute,
         RoleMethod;
 
     /**
@@ -20,4 +23,14 @@ class Role extends SpatieRole
     protected $with = [
         'permissions',
     ];
+
+    /**
+     * Create a new factory instance for the model.
+     *
+     * @return \Illuminate\Database\Eloquent\Factories\Factory
+     */
+    protected static function newFactory()
+    {
+        return RoleFactory::new();
+    }
 }
