@@ -1,10 +1,12 @@
 <?php
 
-namespace App\Http\Livewire;
+namespace App\Http\Livewire\Backend;
 
 use App\Domains\Auth\Models\Role;
+use App\Domains\Auth\Models\User;
 use Illuminate\Database\Eloquent\Builder;
 use Rappasoft\LaravelLivewireTables\TableComponent;
+use Rappasoft\LaravelLivewireTables\Traits\HtmlComponents;
 use Rappasoft\LaravelLivewireTables\Views\Column;
 
 /**
@@ -12,10 +14,20 @@ use Rappasoft\LaravelLivewireTables\Views\Column;
  */
 class RolesTable extends TableComponent
 {
+    use HtmlComponents;
+
     /**
      * @var string
      */
     public $sortField = 'name';
+
+    /**
+     * @var array
+     */
+    protected $options = [
+        'bootstrap.container' => false,
+        'bootstrap.classes.table' => 'table table-striped',
+    ];
 
     /**
      * @return Builder
@@ -32,24 +44,37 @@ class RolesTable extends TableComponent
     public function columns(): array
     {
         return [
-            Column::make(__('Type'))
-                ->view('backend.auth.role.includes.type', 'role')
-                ->sortable(),
-            Column::make(__('Name'))
+            Column::make(__('Type'), 'type')
+                ->sortable()
+                ->format(function (Role $model) {
+                    if ($model->type === User::TYPE_ADMIN) {
+                        return __('Administrator');
+                    }
+
+                    if ($model->type === User::TYPE_USER) {
+                        return __('User');
+                    }
+
+                    return 'N/A';
+                }),
+            Column::make(__('Name'), 'name')
                 ->searchable()
                 ->sortable(),
             Column::make(__('Permissions'), 'permissions_label')
-                ->customAttribute()
-                ->html()
                 ->searchable(function ($builder, $term) {
                     return $builder->orWhereHas('permissions', function ($query) use ($term) {
                         return $query->where('name', 'like', '%'.$term.'%');
                     });
+                })
+                ->format(function (Role $model) {
+                    return $this->html($model->permissions_label);
                 }),
             Column::make(__('Number of Users'), 'users_count')
                 ->sortable(),
             Column::make(__('Actions'))
-                ->view('backend.auth.role.includes.actions'),
+                ->format(function (Role $model) {
+                    return view('backend.auth.role.includes.actions', ['model' => $model]);
+                }),
         ];
     }
 }
