@@ -1,12 +1,33 @@
 <?php
 
+namespace Database\Factories;
+
 use App\Domains\Auth\Models\Role;
 use App\Domains\Auth\Models\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Role::class, function (Faker $faker) {
-    return [
-        'type' => $faker->randomElement([User::TYPE_ADMIN, User::TYPE_USER]),
-        'name' => $faker->word,
-    ];
-});
+/**
+ * Class RoleFactory.
+ */
+class RoleFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Role::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'type' => $this->faker->randomElement([User::TYPE_ADMIN, User::TYPE_USER]),
+            'name' => $this->faker->word,
+        ];
+    }
+}
