@@ -1,78 +1,135 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
+
 use App\Domains\Auth\Models\User;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 use Illuminate\Support\Str;
 
-/*
-|--------------------------------------------------------------------------
-| Model Factories
-|--------------------------------------------------------------------------
-|
-| This directory should contain each of the model factory definitions for
-| your application. Factories provide a convenient way to generate new
-| model instances for testing / seeding your application's database.
-|
-*/
+/**
+ * Class UserFactory.
+ */
+class UserFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = User::class;
 
-$factory->define(User::class, function (Faker $faker) {
-    return [
-        'type' => $faker->randomElement([User::TYPE_ADMIN, User::TYPE_USER]),
-        'name' => $faker->name,
-        'email' => $faker->unique()->safeEmail,
-        'email_verified_at' => now(),
-        'password' => 'secret',
-        'password_changed_at' => null,
-        'remember_token' => Str::random(10),
-        'active' => true,
-    ];
-});
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'type' => $this->faker->randomElement([User::TYPE_ADMIN, User::TYPE_USER]),
+            'name' => $this->faker->name,
+            'email' => $this->faker->unique()->safeEmail,
+            'email_verified_at' => now(),
+            'password' => 'secret',
+            'password_changed_at' => null,
+            'remember_token' => Str::random(10),
+            'active' => true,
+        ];
+    }
 
-$factory->state(User::class, 'admin', function () {
-    return [
-        'type' => User::TYPE_ADMIN,
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function admin()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'type' => User::TYPE_ADMIN,
+            ];
+        });
+    }
 
-$factory->state(User::class, 'user', function () {
-    return [
-        'type' => User::TYPE_USER,
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function user()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'type' => User::TYPE_USER,
+            ];
+        });
+    }
 
-$factory->state(User::class, 'active', function () {
-    return [
-        'active' => true,
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function active()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'active' => true,
+            ];
+        });
+    }
 
-$factory->state(User::class, 'inactive', function () {
-    return [
-        'active' => false,
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function inactive()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'active' => false,
+            ];
+        });
+    }
 
-$factory->state(User::class, 'confirmed', function () {
-    return [
-        'email_verified_at' => now(),
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function confirmed()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'email_verified_at' => now(),
+            ];
+        });
+    }
 
-$factory->state(User::class, 'unconfirmed', function () {
-    return [
-        'email_verified_at' => null,
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function unconfirmed()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'email_verified_at' => null,
+            ];
+        });
+    }
 
-$factory->state(User::class, 'password_expired', function () {
-    return [
-        'password_changed_at' => now()->subYears(5),
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function passwordExpired()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'password_changed_at' => now()->subYears(5),
+            ];
+        });
+    }
 
-$factory->state(User::class, 'deleted', function () {
-    return [
-        'deleted_at' => now(),
-    ];
-});
+    /**
+     * @return UserFactory
+     */
+    public function deleted()
+    {
+        return $this->state(function (array $attributes) {
+            return [
+                'deleted_at' => now(),
+            ];
+        });
+    }
+}
