@@ -12,6 +12,6 @@ return [
     |
     */
 
-    'failed'   => 'Tyto přihlašovací údaje neodpovídají žadnému záznamu.',
+    'failed' => 'Tyto přihlašovací údaje neodpovídají žadnému záznamu.',
     'throttle' => 'Příliš mnoho pokusů o přihlášení. Zkuste to prosím znovu za :seconds vteřin.',
 ];
