@@ -12,9 +12,9 @@ return [
     |
     */
 
-    'reset'     => 'Heslo bylo obnoveno!',
-    'sent'      => 'E-mail s instrukcemi k obnovení hesla byl odeslán!',
+    'reset' => 'Heslo bylo obnoveno!',
+    'sent' => 'E-mail s instrukcemi k obnovení hesla byl odeslán!',
     'throttled' => 'Počkejte prosím a zkuste to znovu.',
-    'token'     => 'Klíč pro obnovu hesla je nesprávný.',
-    'user'      => 'Nepodařilo se najít uživatele s touto e-mailovou adresou.',
+    'token' => 'Klíč pro obnovu hesla je nesprávný.',
+    'user' => 'Nepodařilo se najít uživatele s touto e-mailovou adresou.',
 ];
