@@ -31,7 +31,7 @@ return [
     'boolean' => 'Le champ :attribute doit être vrai ou faux.',
     'confirmed' => 'Le champ de confirmation :attribute ne correspond pas.',
     'date' => "Le champ :attribute n'est pas une date valide.",
-    'date_equals' => 'The :attribute must be a date equal to :date.',
+    'date_equals' => 'Le champ :attribute doit être une date égale à :date.',
     'date_format' => 'Le champ :attribute ne correspond pas au format :format.',
     'different' => 'Les champs :attribute et :other doivent être différents.',
     'digits' => 'Le champ :attribute doit contenir :digits chiffres.',
@@ -39,21 +39,21 @@ return [
     'dimensions' => 'Les dimensions de l\'image :attribute ne sont pas conformes.',
     'distinct' => 'Le champ :attribute doit être une valeur unique.',
     'email' => 'Le champ :attribute doit être une adresse email valide.',
-    'ends_with' => 'The :attribute must end with one of the following: :values',
+    'ends_with' => 'Le champ :attribute doit se terminer par l\'une des valeurs suivantes:  :values',
     'exists' => 'Le champ :attribute n\'existe pas.',
     'file' => 'Le champ :attribute doit être un fichier.',
     'filled' => 'Le champ :attribute est obligatoire.',
     'gt' => [
-        'numeric' => 'The :attribute must be greater than :value.',
-        'file' => 'The :attribute must be greater than :value kilobytes.',
-        'string' => 'The :attribute must be greater than :value characters.',
-        'array' => 'The :attribute must have more than :value items.',
+        'numeric' => 'Le champ :attribute doit être supérieur à :value.',
+        'file' => 'Le champ :attribute doit être supérieur à :value kilo-octets.',
+        'string' => 'Le champ :attribute doit contenir plus de :value caractères.',
+        'array' => 'Le champ :attribute doit contenir plus de :value éléments.',
     ],
     'gte' => [
-        'numeric' => 'The :attribute must be greater than or equal :value.',
-        'file' => 'The :attribute must be greater than or equal :value kilobytes.',
-        'string' => 'The :attribute must be greater than or equal :value characters.',
-        'array' => 'The :attribute must have :value items or more.',
+        'numeric' => 'Le champ :attribute doit être supérieur ou égal à :value.',
+        'file' => 'Le champ :attribute doit être supérieur ou égal à :value kilo-octets.',
+        'string' => 'Le champ :attribute doit contenir au moins :value caractères.',
+        'array' => 'Le champ :attribute doit contenir au moins :value éléments.',
     ],
     'image' => 'Le champ :attribute doit être une image.',
     'in' => 'Le champ :attribute est invalide.',
@@ -64,16 +64,16 @@ return [
     'ipv6' => 'Le champ :attribute doit être une adresse IPv6 valide.',
     'json' => 'Le champ :attribute doit être un document JSON valide.',
     'lt' => [
-        'numeric' => 'The :attribute must be less than :value.',
-        'file' => 'The :attribute must be less than :value kilobytes.',
-        'string' => 'The :attribute must be less than :value characters.',
-        'array' => 'The :attribute must have less than :value items.',
+        'numeric' => 'Le champ :attribute doit être inférieur à :value.',
+        'file' => 'Le champ :attribute doit être inférieur à :value kilo-octets.',
+        'string' => 'Le champ :attribute doit contenir moins de :value caractères.',
+        'array' => 'Le champ :attribute doit contenir moins de :value éléments.',
     ],
     'lte' => [
-        'numeric' => 'The :attribute must be less than or equal :value.',
-        'file' => 'The :attribute must be less than or equal :value kilobytes.',
-        'string' => 'The :attribute must be less than or equal :value characters.',
-        'array' => 'The :attribute must not have more than :value items.',
+        'numeric' => 'Le champ :attribute doit être inférieur ou égal à :value.',
+        'file' => 'Le champ :attribute doit être inférieur ou égal à :value kilo-octets.',
+        'string' => 'Le champ :attribute doit contenir :value caractères maximum.',
+        'array' => 'Le champ :attribute doit contenir :value éléments maximum.',
     ],
     'max' => [
         'numeric' => 'La valeur de :attribute ne peut être supérieure à :max.',
@@ -90,7 +90,7 @@ return [
         'array' => 'Le tableau :attribute doit contenir au moins :min éléments.',
     ],
     'not_in' => "Le champ :attribute sélectionné n'est pas valide.",
-    'not_regex' => 'The :attribute format is invalid.',
+    'not_regex' => 'Le format du champ :attribute n\'est pas valide.',
     'numeric' => 'Le champ :attribute doit contenir un nombre.',
     'present' => 'Le champ :attribute doit être présent.',
     'regex' => 'Le format du champ :attribute est invalide.',
@@ -108,13 +108,13 @@ return [
         'string' => 'Le texte de :attribute doit contenir :size caractères.',
         'array' => 'Le tableau :attribute doit contenir :size éléments.',
     ],
-    'starts_with' => 'The :attribute must start with one of the following: :values',
+    'starts_with' => 'Le champ :attribute doit commencer par l\'une des valeurs suivantes: :values',
     'string' => 'Le champ :attribute doit être une chaîne de caractères.',
     'timezone' => 'Le champ :attribute doit être un fuseau horaire valide.',
     'unique' => 'La valeur du champ :attribute est déjà utilisée.',
     'uploaded' => 'Le fichier du champ :attribute n\'a pu être téléchargé.',
     'url' => 'Le format de \'URL de :attribute n\'est pas valide.',
-    'uuid' => 'The :attribute must be a valid UUID.',
+    'uuid' => 'Le champ :attribute doit contenir un UUID valide.',
 
     /*
     |--------------------------------------------------------------------------
