@@ -24,7 +24,7 @@
         @endif
 
         <x-slot name="body">
-            <livewire:users-table />
+            <livewire:backend.users-table />
         </x-slot>
     </x-backend.card>
 @endsection
