@@ -11,6 +11,6 @@
 |
 */
 
-//Broadcast::channel('App.User.{id}', function ($user, $id) {
+//Broadcast::channel('App.Domains.Auth.Models.User.{id}', function ($user, $id) {
 //    return (int) $user->id === (int) $id;
 //});
